reusability and custom hooks functionality

+++ user_form_react/src/api/useDelete.js
@@ -1,20 +1,20 @@
 import axios from "axios";
-import { useEffect, useState } from "react";
 
-export const useDelete = (props) => {
-  const [status, setStatus] = useState();
+export const useDelete = () => {
   const baseURL = "http://localhost:3001/users";
-  useEffect(() => {
-    deleteUser(props);
-  }, []);
-  const deleteUser = async (id) => {
-    try {
-      // console.log(id);
-      const response = await axios.delete(`${baseURL}/${id}`);
-      setStatus(response.status);
-    } catch (error) {
-      console.log(error);
-    }
+
+  return {
+    mutateAsyncDelete: async (id) => {
+      try {
+        if (!id) {
+          return "";
+        }
+        // console.log(id);
+        const response = await axios.delete(`${baseURL}/${id}`);
+        return response.status;
+      } catch (error) {
+        console.log(error);
+      }
+    },
   };
-  return status;
 };
